fix: split audit source filter helper
